+++ src/app/Message/Event.php
@@ -1,0 +1,16 @@
+<?php
+
+
+
+namespace App\Message;
+
+
+
+readonly class Event
+{
+    public function __construct (public string $type, public array $data) {}
+}
+
+
+
+?>