+++ src/app/Message/Message.php
@@ -1,0 +1,16 @@
+<?php
+
+
+
+namespace App\Message;
+
+
+
+readonly class Message
+{
+    public function __construct (public Target $target, public Event $event) {}
+}
+
+
+
+?>